Move history plugin confpage.
